Add TestAndroidStitch to run, turn on compareExpectedRegions
